simplified scripts added built-in support for workload identity

+++ database/init-script.sql
@@ -274,7 +274,7 @@ CREATE TABLE sli_metrics (
     "entity" VARCHAR(66),
     "entity_id" INT,
     "name" VARCHAR(20),
-    "value" INT,
+    "value" BIGINT,
     "row_id" SERIAL,
     "row_time" TIMESTAMP,
     PRIMARY KEY ("row_id")
